+++ common/src/main/java/net/modgarden/barricade/block/AdvancedBarrierBlock.java
@@ -2,7 +2,7 @@ package net.modgarden.barricade.block;
 
 import com.mojang.serialization.MapCodec;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.BarrierBlock;
 import net.minecraft.world.level.block.EntityBlock;
@@ -10,6 +10,7 @@ import net.minecraft.world.level.block.RenderShape;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.EntityCollisionContext;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import net.modgarden.barricade.block.entity.AdvancedBarrierBlockEntity;
@@ -34,21 +35,14 @@ public class AdvancedBarrierBlock extends BarrierBlock implements EntityBlock {
 
     @Override
     public VoxelShape getCollisionShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context) {
-        if (level.getBlockEntity(pos) instanceof AdvancedBarrierBlockEntity blockEntity) {
-            if (blockEntity.getBlockedEntities() == null || !blockEntity.getBlockedEntities().canPass(context)) {
-                if (blockEntity.getBlockedDirections() != null && blockEntity.getBlockedDirections().doesNotBlock())
-                    return Shapes.empty();
-                else if (blockEntity.getBlockedDirections() == null || blockEntity.getBlockedDirections().blocksAll())
-                    return super.getCollisionShape(state, level, pos, context);
-                else {
-                    Direction direction = blockEntity.getBlockedDirections().blockingDirection(pos, context);
-                    if (direction == null)
-                        return Shapes.empty();
-                    return super.getCollisionShape(state, level, pos, context);
-                }
-            }
+        if (level.getBlockEntity(pos) instanceof AdvancedBarrierBlockEntity blockEntity && context instanceof EntityCollisionContext entityContext) {
+            ServerLevel serverLevel = null;
+            if (level instanceof ServerLevel)
+                serverLevel = (ServerLevel) level;
+            if (!(blockEntity.getData().directions().doesNotBlock() || entityContext.getEntity() != null && blockEntity.getData().test(serverLevel, entityContext.getEntity(), state, pos)) && (blockEntity.getData().directions().blocksAll() || !blockEntity.getData().directions().test(pos, context)))
+                return Shapes.empty();
         }
-        return Shapes.empty();
+        return super.getCollisionShape(state, level, pos, context);
     }
 
     @Nullable