+++ common/src/main/java/net/modgarden/barricade/block/entity/AdvancedBarrierBlockEntity.java
@@ -1,66 +1,57 @@
 package net.modgarden.barricade.block.entity;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Holder;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.component.DataComponentMap;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtOps;
+import net.minecraft.network.chat.Component;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
+import net.minecraft.world.Nameable;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.modgarden.barricade.component.BlockedDirectionsComponent;
-import net.modgarden.barricade.component.BlockedEntitiesComponent;
+import net.modgarden.barricade.data.AdvancedBarrier;
 import net.modgarden.barricade.registry.BarricadeBlockEntityTypes;
 import net.modgarden.barricade.registry.BarricadeComponents;
 import org.jetbrains.annotations.Nullable;
 
-public class AdvancedBarrierBlockEntity extends BlockEntity {
-    @Nullable
-    private BlockedDirectionsComponent blockedDirections;
-    private BlockedEntitiesComponent blockedEntities;
+public class AdvancedBarrierBlockEntity extends BlockEntity implements Nameable {
+    private Holder<AdvancedBarrier> data;
 
     public AdvancedBarrierBlockEntity(BlockPos pos, BlockState blockState) {
         super(BarricadeBlockEntityTypes.ADVANCED_BARRIER, pos, blockState);
+        this.data = Holder.direct(AdvancedBarrier.DEFAULT);
     }
 
-    @Nullable
-    public BlockedDirectionsComponent getBlockedDirections() {
-        return blockedDirections;
-    }
-
-    @Nullable
-    public BlockedEntitiesComponent getBlockedEntities() {
-        return blockedEntities;
+    public AdvancedBarrier getData() {
+        if (data == null || !data.isBound())
+            return AdvancedBarrier.DEFAULT;
+        return data.value();
     }
 
     @Override
     protected void applyImplicitComponents(BlockEntity.DataComponentInput components) {
-        blockedDirections = components.get(BarricadeComponents.BLOCKED_DIRECTIONS);
-        blockedEntities = components.get(BarricadeComponents.BLOCKED_ENTITIES);
+        data = components.get(BarricadeComponents.ADVANCED_BARRIER);
     }
 
     @Override
     protected void collectImplicitComponents(DataComponentMap.Builder components) {
-        components.set(BarricadeComponents.BLOCKED_DIRECTIONS, blockedDirections);
-        components.set(BarricadeComponents.BLOCKED_ENTITIES, blockedEntities);
+        components.set(BarricadeComponents.ADVANCED_BARRIER, data);
     }
 
     @Override
     protected void loadAdditional(CompoundTag tag, HolderLookup.Provider registries) {
-        if (tag.contains("blocked_directions"))
-            blockedDirections = BlockedDirectionsComponent.CODEC.parse(NbtOps.INSTANCE, tag.get("blocked_directions")).getOrThrow();
-        if (tag.contains("blocked_entities"))
-            blockedEntities = BlockedEntitiesComponent.CODEC.parse(registries.createSerializationContext(NbtOps.INSTANCE), tag.get("blocked_entities")).mapOrElse(success -> success, err -> BlockedEntitiesComponent.EMPTY);
+        if (tag.contains("data"))
+            data = AdvancedBarrier.CODEC.parse(registries.createSerializationContext(NbtOps.INSTANCE), tag.get("data")).getOrThrow();
     }
 
     @Override
     protected void saveAdditional(CompoundTag tag, HolderLookup.Provider registries) {
-        if (blockedDirections != null)
-            tag.put("blocked_directions", BlockedDirectionsComponent.CODEC.encodeStart(NbtOps.INSTANCE, blockedDirections).getOrThrow());
-        if (blockedEntities != null)
-            tag.put("blocked_entities", BlockedEntitiesComponent.CODEC.encodeStart(registries.createSerializationContext(NbtOps.INSTANCE), blockedEntities).getOrThrow());
+        if (data != null && data.unwrapKey().isPresent())
+            tag.put("data", AdvancedBarrier.CODEC.encodeStart(registries.createSerializationContext(NbtOps.INSTANCE), data).getOrThrow());
     }
 
     @Nullable
@@ -72,5 +63,12 @@ public class AdvancedBarrierBlockEntity extends BlockEntity {
     @Override
     public CompoundTag getUpdateTag(HolderLookup.Provider provider) {
         return saveWithoutMetadata(provider);
+    }
+
+    @Override
+    public Component getName() {
+        if (data.value().name().isEmpty())
+            return Component.translatable("block.barricade.advanced_barrier");
+        return data.value().name().get();
     }
 }