+++ common/src/main/java/net/modgarden/barricade/client/renderer/block/AdvancedBarrierBlockRenderer.java
@@ -11,22 +11,19 @@ import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.BlockModelRotation;
 import net.minecraft.client.resources.model.Material;
 import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.core.Direction;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Blocks;
 import net.modgarden.barricade.Barricade;
 import net.modgarden.barricade.block.entity.AdvancedBarrierBlockEntity;
 import net.modgarden.barricade.client.BarricadeClient;
 import net.modgarden.barricade.client.model.AdvancedBarrierBlockUnbakedModel;
-import net.modgarden.barricade.client.util.AdvancedBarrierComponents;
+import net.modgarden.barricade.client.util.AdvancedBarrierModelValues;
 import net.modgarden.barricade.registry.BarricadeTags;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class AdvancedBarrierBlockRenderer implements BlockEntityRenderer<AdvancedBarrierBlockEntity> {
-    private static final Map<AdvancedBarrierComponents, BakedModel> MODEL_MAP = new HashMap<>();
+    private static final Map<AdvancedBarrierModelValues, BakedModel> MODEL_MAP = new HashMap<>();
 
     @Override
     @SuppressWarnings("ConstantConditions")
@@ -34,7 +31,7 @@ public class AdvancedBarrierBlockRenderer implements BlockEntityRenderer<Advance
         if (!Barricade.isOperatorModel(Blocks.BARRIER.defaultBlockState()) || !Minecraft.getInstance().player.canUseGameMasterBlocks() || !Minecraft.getInstance().player.isHolding(stack -> stack.is(BarricadeTags.ItemTags.BARRIERS)))
             return;
 
-        AdvancedBarrierComponents components = new AdvancedBarrierComponents(blockEntity.getBlockedEntities(), blockEntity.getBlockedDirections());
+        AdvancedBarrierModelValues components = new AdvancedBarrierModelValues(blockEntity.getData().directions(), blockEntity.getData().icon().orElse(null));
 
         BarricadeClient.getHelper().tessellateBlock(
                 Minecraft.getInstance().level,
@@ -49,18 +46,11 @@ public class AdvancedBarrierBlockRenderer implements BlockEntityRenderer<Advance
                 OverlayTexture.NO_OVERLAY);
     }
 
-    private static BakedModel createModel(AdvancedBarrierComponents components) {
-        String variant = "";
-        if (components.blockedEntities() != null)
-            variant = components.blockedEntities().icon() + "," + String.join(",", components.blockedEntities().entities().stream().map(either -> either.map(tagKey -> "#" + tagKey.location(), holder -> holder.unwrapKey().map(ResourceKey::location).orElse(ResourceLocation.withDefaultNamespace("null")).toString())).toList());
-
-        if (components.blockedDirections() != null && !components.blockedDirections().doesNotBlock()) {
-            variant = String.join(",", components.blockedDirections().directions().stream().map(Direction::getName).toList());
-        }
-        BlockModel blockModel = new AdvancedBarrierBlockUnbakedModel(components.blockedDirections(), components.blockedEntities());
+    private static BakedModel createModel(AdvancedBarrierModelValues values) {
+        BlockModel blockModel = new AdvancedBarrierBlockUnbakedModel(values.directions(), values.icon());
         return blockModel
                 .bake(
-                        BarricadeClient.getModelBakery().new ModelBakerImpl((modelLocation, material) -> material.sprite(), new ModelResourceLocation(Barricade.asResource("advanced_barrier_item_renderer"), variant)),
+                        BarricadeClient.getModelBakery().new ModelBakerImpl((modelLocation, material) -> material.sprite(), new ModelResourceLocation(Barricade.asResource("advanced_barrier_item_renderer"), values.getVariant())),
                         Material::sprite,
                         BlockModelRotation.X0_Y0
                 );