+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changes
-- Renamed /textures/item/
+- Added Creative Only Barrier.
+- Renamed /assets/barricade/textures/item/barricade/entity/ directory to /assets/barricade/textures/item/barricade/icon/.
 
 ## Bugfixes
 - Fixed barriers not turning invisible under certain circumstances when quick swapping items.